Added Player.position parameter to lua script API.
Player.position a table, rather than a vector3, because lua doesn't like vector3s. Player.position[1] is x, Player.position[2] is y, Player.position[3] is z. This sends position in the cell or worldspace that the player is in. It is currently not possible to get a player's current cell or similar, so you'll have to bare that in mind when using this.